% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lepidochroma_palettes.R
\name{lepidochroma_palettes}
\alias{lepidochroma_palettes}
\title{Get information on available palettes}
\usage{
lepidochroma_palettes()
}
\value{
A data frame with 3 variables:
\itemize{
\item name: palette name
\item type: continuous or discrete
\item colourblind_friendly: a logical value indicating if the palette
is colour blind friendly
}
}
\description{
Returns a data frame with the names and information on the palettes
that are available in the package.
}
\examples{
lepidochroma_palettes()

}
